refactor: move variable/functions from run_profiling into seperate file to avoid future repetition
